# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Financial Agent API
  version: 1.0.0
  description: >
    Standard API contract with unified success/error envelopes.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiMeta:
      type: object
      properties:
        correlationId:
          type: string
    ApiError:
      type: object
      required: [type, title, status, detail, instance, code, timestamp]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          enum:
            - BAD_REQUEST
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          nullable: true
...
    GenericSuccess:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ApiMeta'
paths:
...
  /diagnosis/latest:
    get:
      summary: Load latest diagnostic profile for current user
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Diagnostic profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '404':
          description: No profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'